+++ shaders/main_files/gbuffers_hand.glsl
@@ -2,9 +2,17 @@ varying vec2 texcoord;
 varying vec2 lmcoord;
 varying vec4 glcolor;
 
+#ifdef REFLECTIONS_ENABLED
+	varying vec3 normal;
+#endif
+
 #undef SHADOWS_ENABLED
 
 #include "../lib/lighting.glsl"
+
+#if defined BLOOM_ENABLED && defined REFLECTIONS_ENABLED
+	#define BLOOM_AND_REFLECTIONS
+#endif
 
 
 
@@ -41,9 +49,16 @@ void main() {
 		color = vec4(debugOutput, 1.0);
 	#endif
 	gl_FragData[0] = color;
-	#ifdef BLOOM_ENABLED
+	#ifdef BLOOM_AND_REFLECTIONS
+		/* DRAWBUFFERS:024 */
+		gl_FragData[1] = colorForBloom;
+		gl_FragData[2] = vec4(normal, 1.0);
+	#elif defined BLOOM_ENABLED
 		/* DRAWBUFFERS:02 */
 		gl_FragData[1] = colorForBloom;
+	#elif defined REFLECTIONS_ENABLED
+		/* DRAWBUFFERS:04 */
+		gl_FragData[1] = vec4(normal, 1.0);
 	#endif
 }
 
@@ -67,6 +82,8 @@ void main() {
 	
 	
 	glcolor = gl_Color;
+	
+	normal = gl_NormalMatrix * gl_Normal;
 	
 	
 	doPreLighting();